split connect team into its own file
